l3uunq-image-upload-pipeline-memory-optimization-and-async-thumbnail-processing/repository_before/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
celery==5.3.0
redis==5.0.0
psycopg2-binary==2.9.9
